:sparkles: utilities for type-level natural numbers
